Reimplement radios / checkboxes without JS
The current implementation of radios and checkboxes relies on JS. This removes that reliance by changing the markup to allow the use of input pseudoclasses to infer state.

We didn’t go down this route in the first place due to a perceived requirement for a drop-in solution. In practice we probably should have bitten the bullet and made a breaking change, but hindsight is 20:20.

IE8 doesn’t support the required input pseudoclasses, but it also doesn’t support the pseudoelements needed for the new inputs so that’s OK. It does take a slight usability hit in that it no longer gets the focus ring for the whole block, but that feels like a reasonable compromise given its low usage and the reduction in JS use.

None of this has been retested for accessibility yet, and will need a thorough test schedule before it goes live. It’s entirely possible that it will never make it into elements in favour of simply adding it to the new govuk_frontend, but it’s worth recording an implementation here.