Add basic styling for typographic elements
This adds styling to standard typographic elements
without the need to use classes directly on the elements.
This is done scoped under the `govuk-basic-styles` class
which can be added to the `body` or `main` or just a single component.
That makes it backwards-compatible and people can opt in to use them.